[refactor] Relocated vault-filters to app/modules
