+++ bitwarden_license/src/app/app.module.ts
@@ -10,7 +10,6 @@ import { JslibModule } from "jslib-angular/jslib.module";
 
 import { OssRoutingModule } from "src/app/oss-routing.module";
 import { ServicesModule } from "src/app/services/services.module";
-import { VaultFilterModule } from "src/app/vault/vault-filter/vault-filter.module";
 import { WildcardRoutingModule } from "src/app/wildcard-routing.module";
 
 import { AppRoutingModule } from "./app-routing.module";
@@ -23,7 +22,6 @@ import { MaximumVaultTimeoutPolicyComponent } from "./policies/maximum-vault-tim
   imports: [
     OverlayModule,
     JslibModule,
-    VaultFilterModule,
     BrowserAnimationsModule,
     FormsModule,
     ReactiveFormsModule,
